delete dialog added

+++ src/app/components/detail/detail.component.ts
@@ -9,19 +9,20 @@ import { CoffeeService } from '../../services/coffee.service';
 import { MatDialog } from '@angular/material/dialog';
 import { DialogComponent } from '../dialog/dialog.component';
 import { CardComponent } from "../card/card.component";
+import { DialogDeleteComponent } from '../dialog-delete/dialog-delete.component';
 
 @Component({
-    selector: 'app-detail',
-    standalone: true,
-    templateUrl: './detail.component.html',
-    styleUrls: ['./detail.component.scss'],
-    imports: [CommonModule, MatButtonModule, CardComponent, RouterLink]
+  selector: 'app-detail',
+  standalone: true,
+  templateUrl: './detail.component.html',
+  styleUrls: ['./detail.component.scss'],
+  imports: [CommonModule, MatButtonModule, CardComponent, RouterLink]
 })
 export class DetailComponent {
 
   readonly coffeeDetail$: Observable<CoffeeModel> = this._activatedRoute.params.pipe(
     switchMap(params => this._coffeeService.getOne(+params['id']))
-    )
+  )
 
   constructor(private _coffeeService: CoffeeService, private _activatedRoute: ActivatedRoute, private _router: Router, private _matDialog: MatDialog) {
   }
@@ -29,14 +30,24 @@ export class DetailComponent {
   onUpdate(coffee: CoffeeModel) {
     this._matDialog.open(
       DialogComponent, {
-        data: coffee
-      }
+      data: coffee
+    }
     )
   }
 
   onDelete(id: number) {
-    this._coffeeService.delete(id).subscribe(() => {
-      this._router.navigate([''])
-    })
+    const dialogRef = this._matDialog.open(
+      DialogDeleteComponent)
+
+    dialogRef
+      .afterClosed()
+      .subscribe(result => {
+        if (result) {
+          this._coffeeService.delete(id).subscribe(() => {
+            this._router.navigate([''])
+          })
+        }
+      });
   }
 }
+
